fantasy_football/requirements.txt
pyyaml
matplotlib
pandas
numpy
seaborn
